+++ src/main/java/com/graduatioinProject/sensorMonitoring/formerData/electricCurrent/entity/ElectricCurrent.java
@@ -1,0 +1,17 @@
+package com.graduatioinProject.sensorMonitoring.formerData.electricCurrent.entity;
+
+import com.graduatioinProject.sensorMonitoring.formerData.FormerData;
+import lombok.*;
+
+import javax.persistence.*;
+
+@Data
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ElectricCurrent extends FormerData {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+}
