+++ src/main/java/com/graduatioinProject/sensorMonitoring/formerData/temperature/service/TemperatureService.java
@@ -1,20 +1,17 @@
-package com.graduatioinProject.sensorMonitoring.formerData.temperature;
+package com.graduatioinProject.sensorMonitoring.formerData.temperature.service;
 
+import com.graduatioinProject.sensorMonitoring.formerData.temperature.entity.Temperature;
+import com.graduatioinProject.sensorMonitoring.formerData.temperature.repository.TemperatureRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class TemperatureService {
     private final TemperatureRepository temperatureDataRepository;
-
-    public Optional<Temperature> findTemperature(LocalDate date, Long nodePort) {
-        return temperatureDataRepository.findByDateAndNodePort(date, nodePort);
-    }
 
     public List<Temperature> findTemperatureList(LocalDate start, LocalDate end, Long nodePort) {
         return temperatureDataRepository.findAllByDateBetweenAndNodePort(start, end, nodePort);
